refactor networking client
