+++ Sources/WalletConnectNetworking/NetworkingConfig.swift
@@ -1,0 +1,11 @@
+import Foundation
+import WalletConnectRelay
+
+extension Networking {
+    struct Config {
+        let relayHost: String
+        let projectId: String
+        let socketFactory: WebSocketFactory
+        let socketConnectionType: SocketConnectionType
+    }
+}
